requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Scheduling
apscheduler==3.10.4

# iCal parsing
icalendar==5.0.11
recurring-ical-events==2.2.0

# Google OAuth
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# HTTP client
requests==2.31.0
httpx==0.26.0

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Timezone handling
pytz==2024.1

# Retry logic
tenacity==8.2.3
